Added search page
